requirements.txt
request 
beautifulsoup4 

tqdm